+++ src/js/feed.js
@@ -1,8 +1,9 @@
 // @ts-check
-//@deno-types=./19.ts
-import { $, halts, hotkey, internals, makelogger, observe, on, tag, traverse } from "./19.js"
+// @deno-types=./19.ts
+// @deno-types=./43.ts
+import { $, halts, hotkey, makelogger, observe, on, traverse } from "./19.js"
+import { internals, tag, validate } from "./43.js"
 import { AriaBusy } from "./aria.js"
-import { validate } from "./validate.js"
 
 const ilog = makelogger("feed")
